Functional upload

+++ meta.php
@@ -54,13 +54,24 @@ function doUploadImage() {
         return array("status"=>false,"error"=>"No files provided","human_error"=>"Please provide a file to upload");
     }
     $temp = $_FILES["file"]["tmp_name"];
-    $savePath = dirname(__FILE__) . "/species_photos/";
+    $uploadPath = $_REQUEST["uploadpath"];
+    $savePath = dirname(__FILE__) . "/" . $uploadPath;
+    if(!file_exists($savePath)) {
+        return array(
+            "status" => false,
+            "error" => "Bad path '$savePath'",
+            "human_error" => "There is a server misconfiguration preventing your file from being uploaded"
+        );
+    }
     $file = $_FILES["file"]["name"];
     $extension = array_pop(explode(".",$file));
     $newFilePath = md5($file) . "." . $extension;
     $fileWritePath = $savePath . $newFilePath;
+    # We want to suppress the warning on move_uploaded_file, or else
+    # it'll return an invalid JSON response
+    error_reporting(0); # Disable this for debugging
     return array("status"=>move_uploaded_file($temp,$fileWritePath),"original_file"=>$file,"wrote_file"=>$newFilePath,"full_path"=>$fileWritePath);
-    
+
 }
